backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
django-allauth==0.57.0
djangorestframework-simplejwt[crypto]==5.3.0

# File Upload & Storage
django-storages==1.14.2
boto3==1.34.0
Pillow==10.1.0

# API Documentation
drf-spectacular==0.26.5
drf-nested-routers==0.93.4
django-filter==23.5

# Development & Production
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Utilities
celery==5.3.4
redis==5.0.1 